requirements.txt
matplotlib==3.3.0
torch==1.6.0
opencv_python==4.3.0.36
numpy==1.19.1
Pillow==7.2.0
